in progress

+++ myMoney/Record.h
@@ -10,13 +10,13 @@
 
 @interface Record : NSObject
 
-@property (nonatomic, assign) double amount;
-@property (nonatomic, assign) int ID;
+@property (nonatomic, assign) CGFloat amount;
+@property (nonatomic, assign) NSInteger ID;
 @property (nonatomic, strong) NSString *description;
 @property (nonatomic, strong) NSDate *date;
 @property (nonatomic, strong) NSArray *tags;
 
--(id)initWithAmount:(double)amount description:(NSString*)description andTags:(NSArray*)tags;
+-(id)initWithAmount:(CGFloat)amount description:(NSString*)description andTags:(NSArray*)tags;
 
 -(BOOL)isProfit;
